fix: initial login page setup wip
