--- apps/web/components/Form/index.js
@@ -1,1 +1,0 @@
-export * from "./Form";
