+++ apps/web/pages/index.js
@@ -1,6 +1,6 @@
 import Head from "next/head";
 import Image from "next/image";
-import { Form } from "../components/Form";
+import { Login } from "../components/Login";
 import styles from "../styles/Home.module.css";
 import { ChakraProvider, Center } from "@chakra-ui/react";
 
@@ -8,7 +8,7 @@ export default function Home() {
   return (
     <ChakraProvider>
       <Center>
-        <Form />
+        <Login />
       </Center>
     </ChakraProvider>
   );
